pyproject.toml
[project]
name = "code-graph-pilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "neo4j>=6.0.3",
    "networkx>=3.4.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
